backend/requirements.txt
Flask==3.1.0
psycopg2-binary==2.9.10
elasticsearch
minio
python-dotenv